Initial public release
The code is a bit ugly and only basic functionality is there but it has been tested and is fully usable.  I have not provided a schematic but it is fairly self explanatory, just a power transistor to drive the lamp and some configuration switches to select the pattern connected to an Arduino Nano clone.